[apex] Move ApexCSRF rule from security to errorprone
Also move the Helper class into a internal package and
deprecated the old Helper class.

Refs #2064
