+++ pmd-apex/src/main/resources/category/apex/security.xml
@@ -62,30 +62,7 @@ public class Foo {
         </example>
     </rule>
 
-    <rule name="ApexCSRF"
-          since="5.5.3"
-          message="Avoid making DML operations in Apex class constructor/init method"
-          class="net.sourceforge.pmd.lang.apex.rule.security.ApexCSRFRule"
-          externalInfoUrl="${pmd.website.baseurl}/pmd_rules_apex_security.html#apexcsrf">
-        <description>
-Check to avoid making DML operations in Apex class constructor/init method. This prevents
-modification of the database just by accessing a page.
-        </description>
-        <priority>3</priority>
-        <example>
-<![CDATA[
-public class Foo {
-    public init() {
-        insert data;
-    }
-
-    public Foo() {
-        insert data;
-    }
-}
-]]>
-        </example>
-    </rule>
+    <rule name="ApexCSRF" ref="category/apex/errorprone.xml/ApexCSRF" deprecated="true"/>
 
     <rule name="ApexDangerousMethods"
           since="5.5.3"
